list function almost done
