+++ data/meals.csv
@@ -1,0 +1,3 @@
+name, price,
+burger, 1.00,
+kebab, 2.00,
