+++ lib/meal.rb
@@ -1,6 +1,6 @@
 
 class Meal
-  attr_reader: :name, :price
+  attr_reader :name, :price
   def initialize (attributes = {})
     @name = attributes[:name]
     @price = attributes[:price] || 0.to_s
