+++ lib/meals_controller.rb
@@ -1,32 +1,16 @@
 
-require_relative 'meal'
 
 class Meals_Controller
-  def initialize (csvfile)
-    @csvfile = csvfile
-    @meals = []
-    load_csv
+  def initialize(meals_repository)
   end
 
-  def list_meals
-    @meals
+  def list
+    puts meals_repository.list_meals
   end
 
-  def add_meal
-  end
 
-  def delete_meal
-  end
-
-  def update_meal
-  end
 
   private
 
-  def load_csv
-  end
-
-  def save_csv
-  end
 
 end
